requirements.txt
pigpio==1.78
numpy==1.24.2
pygame==2.3.0
smbus2==0.4.2 